refactor: Modularize commands into fgp package
- Create fgp/ package with modular structure:
  - fgp/core/policy.py: Policy evaluation, endpoint mapping
  - fgp/core/graphql.py: GraphQL execution utilities
  - fgp/commands/discussion.py: Discussion command module
  - fgp/commands/sub_issue.py: Sub-issue command module
  - fgp/handler.py: HTTP request handler
  - fgp/server.py: Server entry point

- Simplify main.py to just import and call fgp.server.main()

Adding new features now follows a clear pattern:
1. Create module in fgp/commands/
2. Register in fgp/commands/__init__.py COMMAND_MODULES

Related: #25

✍️ **Author**: Claude Code (DevContainer) with osabe
