+++ fgp/__init__.py
@@ -1,0 +1,7 @@
+"""
+fgp - GitHub Finest-Grained Permission Proxy
+
+A proxy that provides fine-grained permission control for GitHub API access.
+"""
+
+__version__ = "0.2.0"
