+++ main.py
@@ -1,1683 +1,22 @@
 #!/usr/bin/env python3
 """
-GitHub Proxy
+GitHub Finest-Grained Permission Proxy
 
-GitHub API および git smart HTTP protocol への権限制限付きプロキシ。
-Classic PAT をホスト側に置き、AI には許可した操作のみ公開する。
+A proxy that provides fine-grained permission control for GitHub API access.
+Classic PAT is stored on the host side, exposing only permitted operations to AI.
 
-Fine-grained PAT では他ユーザーのリポジトリにアクセスできない制限を回避しつつ、
-必要最小限の権限のみを AI に付与する。
-
-機能:
-- GitHub REST API プロキシ（/repos/... など）
-- git smart HTTP プロキシ（/git/{owner}/{repo}.git/... で clone/fetch/push）
-- AWS IAM 式ポリシー評価（allow/deny ルール）
+Features:
+- GitHub REST API proxy (/repos/... etc.)
+- git smart HTTP proxy (/git/{owner}/{repo}.git/... for clone/fetch/push)
+- AWS IAM-style policy evaluation (allow/deny rules)
 
 Usage:
     python main.py [--port PORT] [--config CONFIG_PATH]
 
-設定ファイルのデフォルトパス: ~/.config/github-proxy/config.json
+Default config path: ~/.config/github-proxy/config.json
 """
 
-import argparse
-import base64
-import fnmatch
-import json
-import re
-import sys
-from pathlib import Path
-from http.server import HTTPServer, BaseHTTPRequestHandler
-from urllib.parse import urlparse
-from urllib.request import Request, urlopen
-from urllib.error import HTTPError, URLError
-
-DEFAULT_CONFIG_PATH = Path.home() / ".config" / "github-proxy" / "config.json"
-DEFAULT_PORT = 8766
-
-# =============================================================================
-# アクション体系とエンドポイントマッピング
-# =============================================================================
-
-# (method, pattern, action) のタプル
-# action は層1 action (pr:xxx) または旧形式 (category:operation)
-# パラメータ分岐が必要なものは _PARAM_BRANCH マーク付き（後で body を見て判定）
-ENDPOINT_ACTIONS = [
-    # metadata:read
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/branches$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/branches/(?P<branch>[^/]+)$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/contributors$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/languages$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/tags$", "metadata:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/topics$", "metadata:read"),
-
-    # actions:read
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/actions/runs$", "actions:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/actions/runs/(?P<run_id>\d+)$", "actions:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/actions/runs/(?P<run_id>\d+)/jobs$", "actions:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/actions/workflows$", "actions:read"),
-
-    # statuses:read
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/commits/(?P<ref>[^/]+)/status$", "statuses:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/commits/(?P<ref>[^/]+)/statuses$", "statuses:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/commits/(?P<ref>[^/]+)/check-runs$", "statuses:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/commits/(?P<ref>[^/]+)/check-suites$", "statuses:read"),
-
-    # code:read
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/contents/(?P<path>.*)$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/refs$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/refs/(?P<ref>.+)$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/commits/(?P<sha>[^/]+)$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/trees/(?P<sha>[^/]+)$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/blobs/(?P<sha>[^/]+)$", "code:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/compare/(?P<basehead>.+)$", "code:read"),
-
-    # code:write
-    ("PUT", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/contents/(?P<path>.*)$", "code:write"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/contents/(?P<path>.*)$", "code:write"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/refs$", "code:write"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/git/refs/(?P<ref>.+)$", "code:write"),
-
-    # issues:read
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues$", "issues:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)$", "issues:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/comments$", "issues:read"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/labels$", "issues:read"),
-
-    # issues:write
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues$", "issues:write"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)$", "issues:write"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/comments$", "issues:write"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/comments/(?P<comment_id>\d+)$", "issues:write"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/labels$", "issues:write"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/labels/(?P<label>[^/]+)$", "issues:write"),
-
-    # =========================================================================
-    # PR 層1 Actions
-    # =========================================================================
-
-    # PR 基本操作 (read)
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls$", "pr:list"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)$", "pr:get"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/commits$", "pr:commits"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/files$", "pr:files"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/merge$", "pr:merge_status"),
-
-    # PR 基本操作 (write) - パラメータ分岐は _PARAM_BRANCH で後で判定
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls$", "pr:create_PARAM_BRANCH"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)$", "pr:update_PARAM_BRANCH"),
-    ("PUT", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/update-branch$", "pr:update_branch"),
-
-    # PR merge - パラメータ分岐
-    ("PUT", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/merge$", "pr:merge_PARAM_BRANCH"),
-
-    # General Comments (Issues API 経由)
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/comments$", "pr:comment_list_all"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/comments$", "pr:comment_list"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/comments/(?P<comment_id>\d+)$", "pr:comment_get"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/(?P<issue_number>\d+)/comments$", "pr:comment_create"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/comments/(?P<comment_id>\d+)$", "pr:comment_update"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/issues/comments/(?P<comment_id>\d+)$", "pr:comment_delete"),
-
-    # Review Comments
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/comments$", "pr:review_comment_list_all"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/comments$", "pr:review_comment_list"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/comments/(?P<comment_id>\d+)$", "pr:review_comment_get"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/comments$", "pr:review_comment_create"),
-    ("PATCH", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/comments/(?P<comment_id>\d+)$", "pr:review_comment_update"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/comments/(?P<comment_id>\d+)$", "pr:review_comment_delete"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/comments/(?P<comment_id>\d+)/replies$", "pr:review_comment_reply"),
-
-    # Reviews
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews$", "pr:review_list"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)$", "pr:review_get"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews$", "pr:review_PARAM_BRANCH"),
-    ("PUT", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)$", "pr:review_update"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)$", "pr:review_delete"),
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)/comments$", "pr:review_comments"),
-    ("PUT", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)/dismissals$", "pr:review_dismiss"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/reviews/(?P<review_id>\d+)/events$", "pr:review_submit_PARAM_BRANCH"),
-
-    # Review Requests
-    ("GET", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/requested_reviewers$", "pr:reviewer_list"),
-    ("POST", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/requested_reviewers$", "pr:reviewer_request"),
-    ("DELETE", r"/repos/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pulls/(?P<pull_number>\d+)/requested_reviewers$", "pr:reviewer_remove"),
-]
-
-# git smart HTTP protocol のエンドポイント
-GIT_ENDPOINT_ACTIONS = [
-    # git:read (clone, fetch)
-    ("GET", r"/git/(?P<owner>[^/]+)/(?P<repo>[^/]+)\.git/info/refs$", "git:read"),  # service=git-upload-pack
-    ("POST", r"/git/(?P<owner>[^/]+)/(?P<repo>[^/]+)\.git/git-upload-pack$", "git:read"),
-    # git:write (push)
-    ("POST", r"/git/(?P<owner>[^/]+)/(?P<repo>[^/]+)\.git/git-receive-pack$", "git:write"),
-]
-
-# GraphQL mutation → action マッピング
-GRAPHQL_MUTATION_ACTIONS = {
-    "addDiscussionComment": "discussions:write",
-}
-
-# =============================================================================
-# 層1 Primitive Actions (PR)
-# =============================================================================
-
-PR_LAYER1_ACTIONS = [
-    # PR 基本操作
-    "pr:list", "pr:get", "pr:create", "pr:create_draft", "pr:update",
-    "pr:close", "pr:reopen", "pr:convert_to_draft", "pr:mark_ready",
-    "pr:commits", "pr:files", "pr:merge_status",
-    "pr:merge_commit", "pr:merge_squash", "pr:merge_rebase",
-    "pr:update_branch",
-    # General Comments
-    "pr:comment_list_all", "pr:comment_list", "pr:comment_get",
-    "pr:comment_create", "pr:comment_update", "pr:comment_delete",
-    # Review Comments
-    "pr:review_comment_list_all", "pr:review_comment_list", "pr:review_comment_get",
-    "pr:review_comment_create", "pr:review_comment_update", "pr:review_comment_delete",
-    "pr:review_comment_reply",
-    # Reviews
-    "pr:review_list", "pr:review_get", "pr:review_pending",
-    "pr:approve", "pr:request_changes", "pr:review_comment_only",
-    "pr:review_update", "pr:review_delete", "pr:review_comments",
-    "pr:review_dismiss",
-    "pr:review_submit_approve", "pr:review_submit_request_changes", "pr:review_submit_comment",
-    # Review Requests
-    "pr:reviewer_list", "pr:reviewer_request", "pr:reviewer_remove",
-]
-
-# =============================================================================
-# 層2 Bundle 定義
-# =============================================================================
-
-# pull-requests:read に含まれる層1 actions
-PULL_REQUESTS_READ_ACTIONS = [
-    "pr:list", "pr:get", "pr:commits", "pr:files", "pr:merge_status",
-    "pr:reviewer_list", "pr:review_list", "pr:review_get", "pr:review_comments",
-    "pr:review_comment_list_all", "pr:review_comment_list", "pr:review_comment_get",
-    "pr:comment_list_all", "pr:comment_list", "pr:comment_get",
-]
-
-# pull-requests:write に追加される層1 actions (read は含まない)
-PULL_REQUESTS_WRITE_ONLY_ACTIONS = [
-    "pr:create", "pr:create_draft", "pr:update", "pr:close", "pr:reopen",
-    "pr:convert_to_draft", "pr:mark_ready", "pr:update_branch",
-    "pr:reviewer_request", "pr:reviewer_remove",
-    "pr:review_pending", "pr:approve", "pr:request_changes", "pr:review_comment_only",
-    "pr:review_update", "pr:review_delete", "pr:review_dismiss",
-    "pr:review_submit_approve", "pr:review_submit_request_changes", "pr:review_submit_comment",
-    "pr:review_comment_create", "pr:review_comment_update", "pr:review_comment_delete",
-    "pr:review_comment_reply",
-    "pr:comment_create", "pr:comment_update", "pr:comment_delete",
-]
-
-# pulls:contribute に追加される層1 actions (read は含まない)
-PULLS_CONTRIBUTE_ONLY_ACTIONS = [
-    "pr:create", "pr:create_draft", "pr:update",
-    "pr:convert_to_draft", "pr:mark_ready",
-    "pr:comment_create", "pr:comment_update",
-    "pr:review_comment_create", "pr:review_comment_update", "pr:review_comment_reply",
-    "pr:review_pending", "pr:approve", "pr:request_changes", "pr:review_comment_only",
-    "pr:review_update",
-    "pr:review_submit_approve", "pr:review_submit_request_changes", "pr:review_submit_comment",
-    "pr:reviewer_request",
-]
-
-# 層2 Bundle → 層1 展開マッピング
-BUNDLE_EXPANSION = {
-    "pull-requests:read": PULL_REQUESTS_READ_ACTIONS,
-    "pull-requests:write": PULL_REQUESTS_READ_ACTIONS + PULL_REQUESTS_WRITE_ONLY_ACTIONS,
-    "pulls:contribute": PULL_REQUESTS_READ_ACTIONS + PULLS_CONTRIBUTE_ONLY_ACTIONS,
-    # CLI 高レベル action → 層1 展開
-    "pr:merge": ["pr:merge_commit", "pr:merge_squash", "pr:merge_rebase"],
-}
-
-# ワイルドカード展開用のアクション一覧
-ALL_ACTIONS = [
-    "metadata:read",
-    "actions:read",
-    "statuses:read",
-    "code:read", "code:write",
-    "issues:read", "issues:write",
-    "git:read", "git:write",
-    "discussions:write",
-    "subissues:list", "subissues:parent", "subissues:add", "subissues:remove", "subissues:reprioritize",
-    # 高レベル PR action（cli_args_to_action 用）
-    "pr:read", "pr:create", "pr:write", "pr:merge", "pr:comment", "pr:review",
-] + PR_LAYER1_ACTIONS
-
-# カテゴリごとのアクション（ワイルドカード展開用）
-ACTION_CATEGORIES = {
-    "metadata": ["metadata:read"],
-    "actions": ["actions:read"],
-    "statuses": ["statuses:read"],
-    "code": ["code:read", "code:write"],
-    "issues": ["issues:read", "issues:write"],
-    "pr": ["pr:read", "pr:create", "pr:write", "pr:merge", "pr:comment", "pr:review"] + PR_LAYER1_ACTIONS,
-    "git": ["git:read", "git:write"],
-    "discussions": ["discussions:read", "discussions:write"],
-    "subissues": ["subissues:list", "subissues:parent", "subissues:add", "subissues:remove", "subissues:reprioritize"],
-}
-
-
-# =============================================================================
-# ポリシー評価
-# =============================================================================
-
-def expand_action_pattern(pattern: str) -> list[str]:
-    """
-    アクションパターンを展開する
-    - "*" → 全アクション
-    - "issues:*" → issues:read, issues:write
-    - "pull-requests:read" → 層1 actions に展開
-    - "pulls:contribute" → 層1 actions に展開
-    - "pr:list" → pr:list (層1 action そのまま)
-    """
-    if pattern == "*":
-        return ALL_ACTIONS.copy()
-
-    # 層2 Bundle 展開
-    if pattern in BUNDLE_EXPANSION:
-        return BUNDLE_EXPANSION[pattern].copy()
-
-    # カテゴリワイルドカード展開
-    if pattern.endswith(":*"):
-        category = pattern[:-2]
-        if category in ACTION_CATEGORIES:
-            return ACTION_CATEGORIES[category].copy()
-        return []
-
-    # 層1 action または旧 action
-    if pattern in ALL_ACTIONS:
-        return [pattern]
-
-    return []
-
-
-def expand_repo_pattern(pattern: str, repo: str) -> bool:
-    """
-    リポジトリパターンがマッチするかチェック（case-insensitive）
-    - "*" → 全リポジトリ
-    - "owner/*" → owner の全リポジトリ
-    - "owner/repo" → 完全一致
-    """
-    pattern_lower = pattern.lower()
-    repo_lower = repo.lower()
-
-    if pattern_lower == "*":
-        return True
-
-    if pattern_lower.endswith("/*"):
-        owner_pattern = pattern_lower[:-2]
-        repo_owner = repo_lower.split("/")[0]
-        return owner_pattern == repo_owner
-
-    return fnmatch.fnmatch(repo_lower, pattern_lower)
-
-
-def evaluate_policy(action: str, repo: str, rules: list[dict]) -> tuple[bool, str]:
-    """
-    AWS IAM 式のポリシー評価
-
-    評価ロジック:
-    1. デフォルト: 暗黙の Deny
-    2. 一つでも Deny にマッチ → 即拒否（Deny always wins）
-    3. 一つでも Allow にマッチ → 許可
-    4. 何もマッチしない → 拒否
-
-    Returns: (allowed, reason)
-    """
-    has_allow = False
-
-    for rule in rules:
-        effect = rule.get("effect", "").lower()
-        actions = rule.get("actions", [])
-        repos = rule.get("repos", [])
-
-        # アクションがマッチするか
-        action_match = False
-        for action_pattern in actions:
-            expanded = expand_action_pattern(action_pattern)
-            if action in expanded:
-                action_match = True
-                break
-
-        if not action_match:
-            continue
-
-        # リポジトリがマッチするか
-        repo_match = False
-        for repo_pattern in repos:
-            if expand_repo_pattern(repo_pattern, repo):
-                repo_match = True
-                break
-
-        if not repo_match:
-            continue
-
-        # 両方マッチした場合
-        if effect == "deny":
-            return False, f"Denied by rule: {rule}"
-        elif effect == "allow":
-            has_allow = True
-
-    if has_allow:
-        return True, "Allowed"
-    else:
-        return False, f"No matching allow rule for {action} on {repo}"
-
-
-# =============================================================================
-# パラメータ分岐
-# =============================================================================
-
-def resolve_param_branch(action: str, body: bytes | None) -> str:
-    """
-    _PARAM_BRANCH マーク付きの action をリクエストボディに基づいて解決する
-
-    パラメータ分岐:
-    - pr:create_PARAM_BRANCH → draft で pr:create or pr:create_draft
-    - pr:update_PARAM_BRANCH → state/draft で pr:update, pr:close, pr:reopen, pr:convert_to_draft, pr:mark_ready
-    - pr:merge_PARAM_BRANCH → merge_method で pr:merge_commit, pr:merge_squash, pr:merge_rebase
-    - pr:review_PARAM_BRANCH → event で pr:review_pending, pr:approve, pr:request_changes, pr:review_comment_only
-    - pr:review_submit_PARAM_BRANCH → event で pr:review_submit_approve, pr:review_submit_request_changes, pr:review_submit_comment
-    """
-    if "_PARAM_BRANCH" not in action:
-        return action
-
-    # ボディをパース
-    body_json = {}
-    if body:
-        try:
-            body_json = json.loads(body.decode("utf-8"))
-        except (json.JSONDecodeError, UnicodeDecodeError):
-            pass
-
-    # pr:create_PARAM_BRANCH
-    if action == "pr:create_PARAM_BRANCH":
-        if body_json.get("draft", False):
-            return "pr:create_draft"
-        return "pr:create"
-
-    # pr:update_PARAM_BRANCH
-    if action == "pr:update_PARAM_BRANCH":
-        state = body_json.get("state")
-        draft = body_json.get("draft")
-
-        if state == "closed":
-            return "pr:close"
-        if state == "open":
-            return "pr:reopen"
-        if draft is True:
-            return "pr:convert_to_draft"
-        if draft is False:
-            return "pr:mark_ready"
-        return "pr:update"
-
-    # pr:merge_PARAM_BRANCH
-    if action == "pr:merge_PARAM_BRANCH":
-        merge_method = body_json.get("merge_method", "merge")
-        if merge_method == "squash":
-            return "pr:merge_squash"
-        if merge_method == "rebase":
-            return "pr:merge_rebase"
-        return "pr:merge_commit"
-
-    # pr:review_PARAM_BRANCH
-    if action == "pr:review_PARAM_BRANCH":
-        event = body_json.get("event", "").upper()
-        if event == "APPROVE":
-            return "pr:approve"
-        if event == "REQUEST_CHANGES":
-            return "pr:request_changes"
-        if event == "COMMENT":
-            return "pr:review_comment_only"
-        return "pr:review_pending"
-
-    # pr:review_submit_PARAM_BRANCH
-    if action == "pr:review_submit_PARAM_BRANCH":
-        event = body_json.get("event", "").upper()
-        if event == "APPROVE":
-            return "pr:review_submit_approve"
-        if event == "REQUEST_CHANGES":
-            return "pr:review_submit_request_changes"
-        return "pr:review_submit_comment"
-
-    # 未知の _PARAM_BRANCH（フォールバック）
-    return action.replace("_PARAM_BRANCH", "")
-
-
-# =============================================================================
-# エンドポイントマッチング
-# =============================================================================
-
-def match_endpoint(method: str, path: str) -> tuple[str | None, dict]:
-    """
-    エンドポイントがマッチするかチェック
-    Returns: (action, captured_groups) or (None, {})
-    """
-    for allowed_method, pattern, action in ENDPOINT_ACTIONS:
-        if method != allowed_method:
-            continue
-        match = re.match(pattern, path)
-        if match:
-            return action, match.groupdict()
-    return None, {}
-
-
-def match_git_endpoint(method: str, path: str, query: str) -> tuple[str | None, dict]:
-    """
-    git smart HTTP エンドポイントにマッチするかチェック
-    info/refs の場合は service パラメータで read/write を判定
-    Returns: (action, captured_groups) or (None, {})
-    """
-    for allowed_method, pattern, action in GIT_ENDPOINT_ACTIONS:
-        if method != allowed_method:
-            continue
-        match = re.match(pattern, path)
-        if match:
-            # info/refs の場合、service パラメータで判定
-            if path.endswith("/info/refs"):
-                if "service=git-receive-pack" in query:
-                    return "git:write", match.groupdict()
-                else:
-                    return "git:read", match.groupdict()
-            return action, match.groupdict()
-    return None, {}
-
-
-# =============================================================================
-# PAT 選択
-# =============================================================================
-
-def select_pat(repo: str, config: dict) -> str:
-    """
-    リポジトリに基づいて適切な PAT を選択する
-
-    1. repo にマッチする fine-grained PAT を探す
-    2. マッチしたらそれを使う
-    3. どれにもマッチしなければ classic PAT（fallback）
-    """
-    for fg_pat in config.get("fine_grained_pats", []):
-        for repo_pattern in fg_pat.get("repos", []):
-            if expand_repo_pattern(repo_pattern, repo):
-                return fg_pat["pat"]
-
-    # fallback to classic PAT
-    return config["classic_pat"]
-
-
-# =============================================================================
-# 設定ファイル
-# =============================================================================
-
-def load_config(config_path: Path) -> dict:
-    """設定ファイルを読み込む"""
-    if not config_path.exists():
-        print(f"Error: Config file not found: {config_path}", file=sys.stderr)
-        print(f"\nCreate the config file with:", file=sys.stderr)
-        print(f"  mkdir -p {config_path.parent}", file=sys.stderr)
-        print(f"  cat > {config_path} << 'EOF'", file=sys.stderr)
-        print("""{
-  "classic_pat": "ghp_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
-  "fine_grained_pats": [
-    { "pat": "github_pat_xxx", "repos": ["owner/*"] }
-  ],
-  "rules": [
-    { "effect": "allow", "actions": ["*"], "repos": ["owner/repo"] },
-    { "effect": "deny", "actions": ["pr:merge_*"], "repos": ["*"] }
-  ]
-}
-EOF""", file=sys.stderr)
-        sys.exit(1)
-
-    # 設定ファイルの権限チェック（600 以外は拒否）
-    stat_info = config_path.stat()
-    if stat_info.st_mode & 0o077:
-        print(f"Error: Config file has insecure permissions: {oct(stat_info.st_mode)[-3:]}", file=sys.stderr)
-        print(f"Run: chmod 600 {config_path}", file=sys.stderr)
-        sys.exit(1)
-
-    try:
-        with open(config_path) as f:
-            config = json.load(f)
-    except json.JSONDecodeError as e:
-        print(f"Error: Invalid JSON in config file: {e}", file=sys.stderr)
-        sys.exit(1)
-
-    # 必須フィールドチェック
-    if "classic_pat" not in config:
-        print("Error: Missing required field: classic_pat", file=sys.stderr)
-        sys.exit(1)
-
-    if "rules" not in config:
-        print("Error: Missing required field: rules", file=sys.stderr)
-        sys.exit(1)
-
-    if not isinstance(config["rules"], list) or len(config["rules"]) == 0:
-        print("Error: rules must be a non-empty list", file=sys.stderr)
-        sys.exit(1)
-
-    # fine_grained_pats のバリデーション（オプショナル）
-    if "fine_grained_pats" in config:
-        if not isinstance(config["fine_grained_pats"], list):
-            print("Error: fine_grained_pats must be a list", file=sys.stderr)
-            sys.exit(1)
-        for i, fg_pat in enumerate(config["fine_grained_pats"]):
-            if "pat" not in fg_pat:
-                print(f"Error: fine_grained_pats[{i}] missing 'pat'", file=sys.stderr)
-                sys.exit(1)
-            if "repos" not in fg_pat or not isinstance(fg_pat["repos"], list):
-                print(f"Error: fine_grained_pats[{i}] missing or invalid 'repos'", file=sys.stderr)
-                sys.exit(1)
-    else:
-        config["fine_grained_pats"] = []
-
-    # ルールのバリデーション
-    for i, rule in enumerate(config["rules"]):
-        if "effect" not in rule:
-            print(f"Error: Rule {i} missing 'effect'", file=sys.stderr)
-            sys.exit(1)
-        if rule["effect"] not in ["allow", "deny"]:
-            print(f"Error: Rule {i} effect must be 'allow' or 'deny'", file=sys.stderr)
-            sys.exit(1)
-        if "actions" not in rule or not isinstance(rule["actions"], list):
-            print(f"Error: Rule {i} missing or invalid 'actions'", file=sys.stderr)
-            sys.exit(1)
-        if "repos" not in rule or not isinstance(rule["repos"], list):
-            print(f"Error: Rule {i} missing or invalid 'repos'", file=sys.stderr)
-            sys.exit(1)
-
-    return config
-
-
-# =============================================================================
-# HTTP ハンドラ
-# =============================================================================
-
-class GitHubProxyHandler(BaseHTTPRequestHandler):
-    """GitHub API および git smart HTTP へのプロキシハンドラ"""
-
-    config: dict = {}
-
-    def log_message(self, format, *args):
-        print(f"[{self.log_date_time_string()}] {format % args}")
-
-    def route_request(self, method: str):
-        """リクエストを適切なハンドラにルーティング"""
-        parsed = urlparse(self.path)
-        path = parsed.path
-
-        if path.startswith("/git/"):
-            self.handle_git_request(method)
-        elif path == "/cli":
-            self.handle_cli_request(method)
-        elif path == "/auth/status":
-            self.handle_auth_status(method)
-        else:
-            self.send_error(404, f"Unknown endpoint: {path}")
-
-    def handle_auth_status(self, method: str):
-        """Check authentication status for all configured PATs"""
-        if method != "GET":
-            self.send_error(405, "Only GET is allowed")
-            return
-
-        result = {
-            "classic_pat": self._check_pat_status(
-                self.config["classic_pat"],
-                pat_type="classic"
-            ),
-            "fine_grained_pats": [],
-        }
-
-        for fg_pat in self.config.get("fine_grained_pats", []):
-            status = self._check_pat_status(
-                fg_pat["pat"],
-                pat_type="fine_grained",
-                repos=fg_pat.get("repos", [])
-            )
-            result["fine_grained_pats"].append(status)
-
-        self.send_response(200)
-        self.send_header("Content-Type", "application/json")
-        self.send_header("Access-Control-Allow-Origin", "*")
-        self.end_headers()
-        self.wfile.write(json.dumps(result, indent=2).encode("utf-8"))
-
-    def _check_pat_status(
-        self, pat: str, pat_type: str, repos: list[str] | None = None
-    ) -> dict:
-        """Validate a PAT by calling GitHub API /user endpoint"""
-        # Mask the PAT for display (show first 4 and last 4 chars)
-        if len(pat) > 12:
-            masked = f"{pat[:4]}...{pat[-4:]}"
-        else:
-            masked = "****"
-
-        try:
-            req = Request(
-                "https://api.github.com/user",
-                headers={
-                    "Authorization": f"Bearer {pat}",
-                    "Accept": "application/vnd.github+json",
-                    "User-Agent": "fgp-proxy",
-                    "X-GitHub-Api-Version": "2022-11-28",
-                },
-            )
-            with urlopen(req, timeout=10) as resp:
-                user_data = json.loads(resp.read().decode("utf-8"))
-                scopes = resp.headers.get("X-OAuth-Scopes", "")
-
-                result = {
-                    "valid": True,
-                    "masked_token": masked,
-                    "user": user_data.get("login"),
-                    "type": pat_type,
-                }
-
-                if pat_type == "classic":
-                    result["scopes"] = [s.strip() for s in scopes.split(",") if s.strip()]
-                else:
-                    # Fine-grained PAT
-                    result["repos"] = repos or []
-                    # Note: Fine-grained PAT expiration is not available via API
-
-                return result
-
-        except HTTPError as e:
-            return {
-                "valid": False,
-                "masked_token": masked,
-                "type": pat_type,
-                "error": f"HTTP {e.code}: {e.reason}",
-                "repos": repos if pat_type == "fine_grained" else None,
-            }
-        except URLError as e:
-            return {
-                "valid": False,
-                "masked_token": masked,
-                "type": pat_type,
-                "error": str(e.reason),
-                "repos": repos if pat_type == "fine_grained" else None,
-            }
-        except Exception as e:
-            return {
-                "valid": False,
-                "masked_token": masked,
-                "type": pat_type,
-                "error": str(e),
-                "repos": repos if pat_type == "fine_grained" else None,
-            }
-
-    def handle_cli_request(self, method: str):
-        """CLI コマンドを受け取って gh を実行する（fgh 用）"""
-        if method != "POST":
-            self.send_error(405, "Only POST is allowed")
-            return
-
-        # リクエストボディの読み取り
-        content_length = int(self.headers.get("Content-Length", 0))
-        if content_length == 0:
-            self.send_error(400, "Request body required")
-            return
-
-        body = self.rfile.read(content_length)
-
-        try:
-            data = json.loads(body.decode("utf-8"))
-        except json.JSONDecodeError:
-            self.send_error(400, "Invalid JSON in request body")
-            return
-
-        args = data.get("args", [])
-        repo = data.get("repo")
-
-        if not args:
-            self.send_error(400, "args is required")
-            return
-
-        if not repo:
-            self.send_error(400, "repo is required")
-            return
-
-        # コマンドから action を判定
-        action, error_message = self.cli_args_to_action(args)
-        if action is None:
-            if error_message:
-                self.send_error(403, error_message)
-            else:
-                self.send_error(403, f"Unknown or unsupported command: {args}")
-            return
-
-        # ポリシー評価
-        allowed, reason = evaluate_policy(action, repo, self.config["rules"])
-        if not allowed:
-            self.send_error(403, reason)
-            return
-
-        # PAT 選択
-        pat = select_pat(repo, self.config)
-
-        # コマンド実行
-        try:
-            cmd = args[0]
-            owner, repo_name = repo.split("/")
-
-            # sub-issue コマンドは GraphQL で実行（gh にはないカスタムコマンド）
-            if cmd == "sub-issue":
-                result = self.execute_sub_issue_cli(args[1:], owner, repo_name, pat)
-            # discussion コマンドも GraphQL で実行
-            elif cmd == "discussion":
-                result = self.execute_discussion_cli(args[1:], owner, repo_name, pat)
-            else:
-                # 標準の gh コマンドは subprocess で実行
-                result = self.execute_gh_cli(args, repo, pat)
-
-            self.send_response(200)
-            self.send_header("Content-Type", "application/json")
-            self.send_header("Access-Control-Allow-Origin", "*")
-            self.end_headers()
-            self.wfile.write(json.dumps(result).encode("utf-8"))
-
-        except ValueError as e:
-            self.send_error(400, str(e))
-        except Exception as e:
-            self.send_error(500, str(e))
-
-    def execute_gh_cli(self, args: list[str], repo: str, pat: str) -> dict:
-        """標準の gh コマンドを subprocess で実行"""
-        import subprocess
-        import os
-
-        gh_args = ["gh"] + args + ["-R", repo]
-
-        result = subprocess.run(
-            gh_args,
-            capture_output=True,
-            text=True,
-            timeout=60,
-            env={
-                **os.environ,
-                "GH_TOKEN": pat,
-                "GH_HOST": "github.com",
-                "GH_FORCE_TTY": "1",
-                "NO_COLOR": "1",
-            }
-        )
-
-        return {
-            "exit_code": result.returncode,
-            "stdout": result.stdout,
-            "stderr": result.stderr
-        }
-
-    def execute_sub_issue_cli(self, args: list[str], owner: str, repo: str, pat: str) -> dict:
-        """sub-issue コマンドを GraphQL で実行"""
-        if not args:
-            raise ValueError("sub-issue subcommand required")
-
-        subcmd = args[0]
-        rest = args[1:]
-
-        if subcmd == "list":
-            if not rest:
-                raise ValueError("issue number required")
-            issue_number = int(rest[0])
-            result = self.execute_list_sub_issues(owner, repo, issue_number, pat)
-            # stdout 形式に変換
-            lines = []
-            for item in result.get("sub_issues", []):
-                lines.append(f"{item['number']}\t{item['state']}\t{item['title']}")
-            return {"exit_code": 0, "stdout": "\n".join(lines), "stderr": ""}
-
-        elif subcmd == "parent":
-            if not rest:
-                raise ValueError("issue number required")
-            issue_number = int(rest[0])
-            result = self.execute_get_parent_issue(owner, repo, issue_number, pat)
-            parent = result.get("parent")
-            if parent:
-                stdout = f"{parent['number']}\t{parent['state']}\t{parent['title']}"
-            else:
-                stdout = "No parent issue"
-            return {"exit_code": 0, "stdout": stdout, "stderr": ""}
-
-        elif subcmd == "add":
-            if len(rest) < 2:
-                raise ValueError("parent and child issue numbers required")
-            parent_number = int(rest[0])
-            child_number = int(rest[1])
-            self.execute_add_sub_issue(owner, repo, parent_number, child_number, pat)
-            return {"exit_code": 0, "stdout": f"Added #{child_number} as sub-issue of #{parent_number}", "stderr": ""}
-
-        elif subcmd == "remove":
-            if len(rest) < 2:
-                raise ValueError("parent and child issue numbers required")
-            parent_number = int(rest[0])
-            child_number = int(rest[1])
-            self.execute_remove_sub_issue(owner, repo, parent_number, child_number, pat)
-            return {"exit_code": 0, "stdout": f"Removed #{child_number} from #{parent_number}", "stderr": ""}
-
-        elif subcmd == "reorder":
-            if len(rest) < 2:
-                raise ValueError("parent and child issue numbers required")
-            parent_number = int(rest[0])
-            child_number = int(rest[1])
-            # --before/--after を探す
-            before_number = None
-            after_number = None
-            i = 2
-            while i < len(rest):
-                if rest[i] == "--before" and i + 1 < len(rest):
-                    before_number = int(rest[i + 1])
-                    i += 2
-                elif rest[i] == "--after" and i + 1 < len(rest):
-                    after_number = int(rest[i + 1])
-                    i += 2
-                else:
-                    i += 1
-            if not before_number and not after_number:
-                raise ValueError("--before or --after required")
-            self.execute_reprioritize_sub_issue(owner, repo, parent_number, child_number, before_number, after_number, pat)
-            return {"exit_code": 0, "stdout": "Reordered", "stderr": ""}
-
-        else:
-            raise ValueError(f"Unknown sub-issue subcommand: {subcmd}")
-
-    def cli_args_to_action(self, args: list[str]) -> tuple[str | None, str | None]:
-        """
-        CLI 引数から action を判定
-
-        Returns:
-            (action, error_message)
-            - action が判定できた場合: (action, None)
-            - 明示的に禁止する場合: (None, error_message)
-            - 不明なコマンドの場合: (None, None)
-        """
-        if len(args) < 1:
-            return None, None
-
-        cmd = args[0]
-        subcmd = args[1] if len(args) > 1 else None
-
-        # api コマンド
-        if cmd == "api":
-            if subcmd == "graphql":
-                return None, "GraphQL API is not allowed via proxy. Use high-level commands (issue, pr, sub-issue) instead."
-            # REST API は後で対応、今は禁止
-            return None, "Direct API calls are not supported yet. Use high-level commands (issue, pr, sub-issue) instead."
-
-        if subcmd is None:
-            return None, None
-
-        # sub-issue コマンド
-        if cmd == "sub-issue":
-            action_map = {
-                "list": "subissues:list",
-                "parent": "subissues:parent",
-                "add": "subissues:add",
-                "remove": "subissues:remove",
-                "reorder": "subissues:reprioritize",
-            }
-            action = action_map.get(subcmd)
-            return (action, None) if action else (None, None)
-
-        # discussion コマンド
-        if cmd == "discussion":
-            if subcmd in ["list", "view"]:
-                return "discussions:read", None
-            elif subcmd in ["create", "edit", "comment"]:
-                return "discussions:write", None
-
-        # issue コマンド
-        if cmd == "issue":
-            if subcmd in ["list", "view"]:
-                return "issues:read", None
-            elif subcmd in ["create", "edit", "close", "reopen", "comment"]:
-                return "issues:write", None
-
-        # pr コマンド
-        if cmd == "pr":
-            if subcmd in ["list", "view", "diff", "checks"]:
-                return "pr:read", None
-            elif subcmd in ["create"]:
-                return "pr:create", None
-            elif subcmd in ["edit", "close", "reopen"]:
-                return "pr:write", None
-            elif subcmd == "merge":
-                # Layer 1: detect merge method from args
-                if "--squash" in args or "-s" in args:
-                    return "pr:merge_squash", None
-                elif "--rebase" in args or "-r" in args:
-                    return "pr:merge_rebase", None
-                elif "--merge" in args or "-m" in args:
-                    return "pr:merge_commit", None
-                # No option specified - gh will error in non-interactive mode
-                # Return layer 2 action which will be expanded during evaluation
-                return "pr:merge", None
-            elif subcmd in ["comment"]:
-                return "pr:comment", None
-            elif subcmd in ["review"]:
-                return "pr:review", None
-
-        return None, None
-
-    def handle_git_request(self, method: str):
-        """git smart HTTP protocol のリクエスト処理"""
-        parsed = urlparse(self.path)
-        path = parsed.path
-        query = parsed.query
-
-        # エンドポイントマッチング
-        action, groups = match_git_endpoint(method, path, query)
-        if action is None:
-            self.send_error(403, f"Git endpoint not allowed: {method} {path}")
-            return
-
-        # リポジトリ取得
-        owner = groups.get("owner")
-        repo = groups.get("repo")
-        if not owner or not repo:
-            self.send_error(400, "Could not determine repository")
-            return
-
-        full_repo = f"{owner}/{repo}"
-
-        # ポリシー評価
-        allowed, reason = evaluate_policy(action, full_repo, self.config["rules"])
-        if not allowed:
-            self.send_error(403, reason)
-            return
-
-        # リクエストボディの読み取り
-        content_length = int(self.headers.get("Content-Length", 0))
-        body = self.rfile.read(content_length) if content_length > 0 else None
-
-        # GitHub にプロキシ
-        try:
-            response_data, response_headers, status_code = self.proxy_git_to_github(
-                method, owner, repo, path, query, body
-            )
-            self.send_response(status_code)
-            if "Content-Type" in response_headers:
-                self.send_header("Content-Type", response_headers["Content-Type"])
-            if "Cache-Control" in response_headers:
-                self.send_header("Cache-Control", response_headers["Cache-Control"])
-            self.send_header("Access-Control-Allow-Origin", "*")
-            self.end_headers()
-            self.wfile.write(response_data)
-        except HTTPError as e:
-            error_body = e.read().decode("utf-8", errors="replace")
-            self.send_error(e.code, f"{e.reason}: {error_body[:200]}")
-        except URLError as e:
-            self.send_error(502, f"Failed to connect to GitHub: {e.reason}")
-        except Exception as e:
-            self.send_error(500, str(e))
-
-    def proxy_git_to_github(
-        self, method: str, owner: str, repo: str, path: str, query: str, body: bytes | None
-    ) -> tuple[bytes, dict, int]:
-        """git smart HTTP を GitHub にプロキシ"""
-        git_path = path.replace(f"/git/{owner}/{repo}.git", f"/{owner}/{repo}.git")
-        url = f"https://github.com{git_path}"
-        if query:
-            url += f"?{query}"
-
-        pat = self.config["classic_pat"]
-        credentials = base64.b64encode(f"x-access-token:{pat}".encode()).decode()
-
-        headers = {
-            "Authorization": f"Basic {credentials}",
-            "User-Agent": "git/2.40.0",
-        }
-
-        if self.headers.get("Content-Type"):
-            headers["Content-Type"] = self.headers.get("Content-Type")
-
-        if self.headers.get("Accept"):
-            headers["Accept"] = self.headers.get("Accept")
-
-        req = Request(url, data=body, headers=headers, method=method)
-
-        with urlopen(req, timeout=60) as response:
-            response_headers = {k: v for k, v in response.headers.items()}
-            return response.read(), response_headers, response.status
-
-    # =========================================================================
-    # Sub-Issues GraphQL Operations (internal use by /cli)
-    # =========================================================================
-
-    def get_issue_node_id(self, owner: str, repo: str, issue_number: int, pat: str) -> str:
-        """Issue の Node ID を取得"""
-        query = """
-        query($owner: String!, $repo: String!, $number: Int!) {
-            repository(owner: $owner, name: $repo) {
-                issue(number: $number) {
-                    id
-                }
-            }
-        }
-        """
-        variables = {"owner": owner, "repo": repo, "number": issue_number}
-        result = self.execute_graphql(query, variables, pat)
-
-        issue = result.get("data", {}).get("repository", {}).get("issue")
-        if not issue:
-            raise ValueError(f"Issue #{issue_number} not found in {owner}/{repo}")
-        return issue["id"]
-
-    def execute_graphql(self, query: str, variables: dict, pat: str) -> dict:
-        """GraphQL クエリを実行"""
-        url = "https://api.github.com/graphql"
-
-        body = {"query": query}
-        if variables:
-            body["variables"] = variables
-
-        headers = {
-            "Authorization": f"bearer {pat}",
-            "Content-Type": "application/json",
-            "User-Agent": "github-proxy/1.0",
-            "GraphQL-Features": "sub_issues",
-        }
-
-        req = Request(url, data=json.dumps(body).encode("utf-8"), headers=headers, method="POST")
-
-        with urlopen(req, timeout=30) as response:
-            result = json.loads(response.read().decode("utf-8"))
-            if "errors" in result:
-                raise ValueError(f"GraphQL error: {result['errors']}")
-            return result
-
-    def execute_list_sub_issues(self, owner: str, repo: str, issue_number: int, pat: str) -> dict:
-        """Issue の sub-issues 一覧を取得"""
-        query = """
-        query($owner: String!, $repo: String!, $number: Int!) {
-            repository(owner: $owner, name: $repo) {
-                issue(number: $number) {
-                    subIssues(first: 50) {
-                        nodes {
-                            number
-                            title
-                            state
-                        }
-                    }
-                }
-            }
-        }
-        """
-        variables = {"owner": owner, "repo": repo, "number": issue_number}
-        result = self.execute_graphql(query, variables, pat)
-
-        issue = result.get("data", {}).get("repository", {}).get("issue")
-        if not issue:
-            raise ValueError(f"Issue #{issue_number} not found in {owner}/{repo}")
-
-        sub_issues = issue.get("subIssues", {}).get("nodes", [])
-        return {"sub_issues": sub_issues}
-
-    def execute_get_parent_issue(self, owner: str, repo: str, issue_number: int, pat: str) -> dict:
-        """Issue の親 issue を取得"""
-        query = """
-        query($owner: String!, $repo: String!, $number: Int!) {
-            repository(owner: $owner, name: $repo) {
-                issue(number: $number) {
-                    parent {
-                        number
-                        title
-                        state
-                    }
-                }
-            }
-        }
-        """
-        variables = {"owner": owner, "repo": repo, "number": issue_number}
-        result = self.execute_graphql(query, variables, pat)
-
-        issue = result.get("data", {}).get("repository", {}).get("issue")
-        if not issue:
-            raise ValueError(f"Issue #{issue_number} not found in {owner}/{repo}")
-
-        parent = issue.get("parent")
-        return {"parent": parent}
-
-    def execute_add_sub_issue(
-        self, owner: str, repo: str, issue_number: int, sub_issue_number: int, pat: str,
-        replace_parent: bool = False
-    ) -> dict:
-        """addSubIssue mutation を実行"""
-        issue_id = self.get_issue_node_id(owner, repo, issue_number, pat)
-        sub_issue_id = self.get_issue_node_id(owner, repo, sub_issue_number, pat)
-
-        mutation = """
-        mutation($issueId: ID!, $subIssueId: ID!, $replaceParent: Boolean) {
-            addSubIssue(input: {issueId: $issueId, subIssueId: $subIssueId, replaceParent: $replaceParent}) {
-                issue { number }
-                subIssue { number }
-            }
-        }
-        """
-        variables = {
-            "issueId": issue_id,
-            "subIssueId": sub_issue_id,
-            "replaceParent": replace_parent
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-
-        return {
-            "success": True,
-            "issue_number": issue_number,
-            "sub_issue_number": sub_issue_number
-        }
-
-    def execute_remove_sub_issue(
-        self, owner: str, repo: str, issue_number: int, sub_issue_number: int, pat: str
-    ) -> dict:
-        """removeSubIssue mutation を実行"""
-        issue_id = self.get_issue_node_id(owner, repo, issue_number, pat)
-        sub_issue_id = self.get_issue_node_id(owner, repo, sub_issue_number, pat)
-
-        mutation = """
-        mutation($issueId: ID!, $subIssueId: ID!) {
-            removeSubIssue(input: {issueId: $issueId, subIssueId: $subIssueId}) {
-                issue { number }
-                subIssue { number }
-            }
-        }
-        """
-        variables = {"issueId": issue_id, "subIssueId": sub_issue_id}
-        result = self.execute_graphql(mutation, variables, pat)
-
-        return {
-            "success": True,
-            "issue_number": issue_number,
-            "sub_issue_number": sub_issue_number
-        }
-
-    def execute_reprioritize_sub_issue(
-        self, owner: str, repo: str, issue_number: int, sub_issue_number: int,
-        before_number: int, after_number: int, pat: str
-    ) -> dict:
-        """reprioritizeSubIssue mutation を実行"""
-        issue_id = self.get_issue_node_id(owner, repo, issue_number, pat)
-        sub_issue_id = self.get_issue_node_id(owner, repo, sub_issue_number, pat)
-
-        before_id = None
-        after_id = None
-        if before_number:
-            before_id = self.get_issue_node_id(owner, repo, before_number, pat)
-        if after_number:
-            after_id = self.get_issue_node_id(owner, repo, after_number, pat)
-
-        mutation = """
-        mutation($issueId: ID!, $subIssueId: ID!, $beforeId: ID, $afterId: ID) {
-            reprioritizeSubIssue(input: {issueId: $issueId, subIssueId: $subIssueId, beforeId: $beforeId, afterId: $afterId}) {
-                issue { number }
-            }
-        }
-        """
-        variables = {
-            "issueId": issue_id,
-            "subIssueId": sub_issue_id,
-            "beforeId": before_id,
-            "afterId": after_id
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-
-        return {
-            "success": True,
-            "issue_number": issue_number,
-            "sub_issue_number": sub_issue_number,
-            "before_number": before_number,
-            "after_number": after_number
-        }
-
-    # =========================================================================
-    # Discussion GraphQL Operations
-    # =========================================================================
-
-    def execute_discussion_cli(
-        self, args: list[str], owner: str, repo: str, pat: str
-    ) -> dict:
-        """discussion コマンドを GraphQL で実行"""
-        if not args:
-            raise ValueError("discussion subcommand required")
-
-        subcmd = args[0]
-        rest = args[1:]
-
-        if subcmd == "list":
-            return self.execute_list_discussions(owner, repo, pat)
-
-        elif subcmd == "view":
-            if not rest:
-                raise ValueError("discussion number required")
-            number = int(rest[0])
-            return self.execute_view_discussion(owner, repo, number, pat)
-
-        elif subcmd == "create":
-            # Parse --title, --body, --category
-            title = None
-            body = None
-            category = None
-            i = 0
-            while i < len(rest):
-                if rest[i] in ["--title", "-t"] and i + 1 < len(rest):
-                    title = rest[i + 1]
-                    i += 2
-                elif rest[i] in ["--body", "-b"] and i + 1 < len(rest):
-                    body = rest[i + 1]
-                    i += 2
-                elif rest[i] in ["--category", "-c"] and i + 1 < len(rest):
-                    category = rest[i + 1]
-                    i += 2
-                else:
-                    i += 1
-            if not title:
-                raise ValueError("--title is required")
-            if not body:
-                raise ValueError("--body is required")
-            if not category:
-                raise ValueError("--category is required")
-            return self.execute_create_discussion(owner, repo, title, body, category, pat)
-
-        elif subcmd == "edit":
-            if not rest:
-                raise ValueError("discussion number required")
-            number = int(rest[0])
-            # Parse --title, --body
-            title = None
-            body = None
-            i = 1
-            while i < len(rest):
-                if rest[i] in ["--title", "-t"] and i + 1 < len(rest):
-                    title = rest[i + 1]
-                    i += 2
-                elif rest[i] in ["--body", "-b"] and i + 1 < len(rest):
-                    body = rest[i + 1]
-                    i += 2
-                else:
-                    i += 1
-            if not title and not body:
-                raise ValueError("--title or --body is required")
-            return self.execute_update_discussion(owner, repo, number, title, body, pat)
-
-        elif subcmd == "comment":
-            if not rest:
-                raise ValueError("discussion number or 'edit' required")
-            # Check if it's "comment edit <comment_id>"
-            if rest[0] == "edit":
-                if len(rest) < 2:
-                    raise ValueError("comment_id required")
-                comment_id = rest[1]
-                # Parse --body
-                body = None
-                i = 2
-                while i < len(rest):
-                    if rest[i] in ["--body", "-b"] and i + 1 < len(rest):
-                        body = rest[i + 1]
-                        i += 2
-                    else:
-                        i += 1
-                if not body:
-                    raise ValueError("--body is required")
-                return self.execute_update_discussion_comment(comment_id, body, pat)
-            else:
-                # Add comment: comment <number> --body "..."
-                number = int(rest[0])
-                # Parse --body, --reply-to
-                body = None
-                reply_to = None
-                i = 1
-                while i < len(rest):
-                    if rest[i] in ["--body", "-b"] and i + 1 < len(rest):
-                        body = rest[i + 1]
-                        i += 2
-                    elif rest[i] == "--reply-to" and i + 1 < len(rest):
-                        reply_to = rest[i + 1]
-                        i += 2
-                    else:
-                        i += 1
-                if not body:
-                    raise ValueError("--body is required")
-                return self.execute_add_discussion_comment(owner, repo, number, body, reply_to, pat)
-
-        else:
-            raise ValueError(f"Unknown discussion subcommand: {subcmd}")
-
-    def get_repository_id(self, owner: str, repo: str, pat: str) -> str:
-        """リポジトリの node ID を取得"""
-        query = """
-        query($owner: String!, $repo: String!) {
-            repository(owner: $owner, name: $repo) {
-                id
-            }
-        }
-        """
-        result = self.execute_graphql(query, {"owner": owner, "repo": repo}, pat)
-        return result["data"]["repository"]["id"]
-
-    def get_discussion_category_id(
-        self, owner: str, repo: str, category_name: str, pat: str
-    ) -> str:
-        """Discussion カテゴリの node ID を取得"""
-        query = """
-        query($owner: String!, $repo: String!) {
-            repository(owner: $owner, name: $repo) {
-                discussionCategories(first: 100) {
-                    nodes {
-                        id
-                        name
-                        slug
-                    }
-                }
-            }
-        }
-        """
-        result = self.execute_graphql(query, {"owner": owner, "repo": repo}, pat)
-        categories = result["data"]["repository"]["discussionCategories"]["nodes"]
-        for cat in categories:
-            if cat["name"].lower() == category_name.lower() or cat["slug"].lower() == category_name.lower():
-                return cat["id"]
-        available = [c["name"] for c in categories]
-        raise ValueError(f"Category '{category_name}' not found. Available: {available}")
-
-    def get_discussion_node_id(
-        self, owner: str, repo: str, number: int, pat: str
-    ) -> str:
-        """Discussion の node ID を取得"""
-        query = """
-        query($owner: String!, $repo: String!, $number: Int!) {
-            repository(owner: $owner, name: $repo) {
-                discussion(number: $number) {
-                    id
-                }
-            }
-        }
-        """
-        result = self.execute_graphql(
-            query, {"owner": owner, "repo": repo, "number": number}, pat
-        )
-        discussion = result["data"]["repository"]["discussion"]
-        if not discussion:
-            raise ValueError(f"Discussion #{number} not found")
-        return discussion["id"]
-
-    def execute_list_discussions(
-        self, owner: str, repo: str, pat: str
-    ) -> dict:
-        """Discussion 一覧を取得"""
-        query = """
-        query($owner: String!, $repo: String!) {
-            repository(owner: $owner, name: $repo) {
-                discussions(first: 30, orderBy: {field: CREATED_AT, direction: DESC}) {
-                    nodes {
-                        number
-                        title
-                        author { login }
-                        createdAt
-                        category { name }
-                        comments { totalCount }
-                    }
-                }
-            }
-        }
-        """
-        result = self.execute_graphql(query, {"owner": owner, "repo": repo}, pat)
-        discussions = result["data"]["repository"]["discussions"]["nodes"]
-
-        # Format output similar to gh issue list
-        lines = []
-        for d in discussions:
-            author = d["author"]["login"] if d["author"] else "ghost"
-            comments = d["comments"]["totalCount"]
-            category = d["category"]["name"] if d["category"] else ""
-            lines.append(f"#{d['number']}\t{d['title']}\t{author}\t{category}\t{comments} comments")
-
-        return {"exit_code": 0, "stdout": "\n".join(lines), "stderr": ""}
-
-    def execute_view_discussion(
-        self, owner: str, repo: str, number: int, pat: str
-    ) -> dict:
-        """Discussion 詳細を取得"""
-        query = """
-        query($owner: String!, $repo: String!, $number: Int!) {
-            repository(owner: $owner, name: $repo) {
-                discussion(number: $number) {
-                    number
-                    title
-                    body
-                    author { login }
-                    createdAt
-                    category { name }
-                    url
-                    comments(first: 50) {
-                        nodes {
-                            id
-                            author { login }
-                            body
-                            createdAt
-                        }
-                    }
-                }
-            }
-        }
-        """
-        result = self.execute_graphql(
-            query, {"owner": owner, "repo": repo, "number": number}, pat
-        )
-        d = result["data"]["repository"]["discussion"]
-        if not d:
-            raise ValueError(f"Discussion #{number} not found")
-
-        author = d["author"]["login"] if d["author"] else "ghost"
-        lines = [
-            f"title:\t{d['title']}",
-            f"number:\t{d['number']}",
-            f"author:\t{author}",
-            f"category:\t{d['category']['name'] if d['category'] else ''}",
-            f"url:\t{d['url']}",
-            f"created:\t{d['createdAt']}",
-            "",
-            "--- BODY ---",
-            d["body"] or "(empty)",
-            "",
-            "--- COMMENTS ---",
-        ]
-        for c in d["comments"]["nodes"]:
-            c_author = c["author"]["login"] if c["author"] else "ghost"
-            lines.append(f"\n[{c['id']}] {c_author} at {c['createdAt']}:")
-            lines.append(c["body"])
-
-        return {"exit_code": 0, "stdout": "\n".join(lines), "stderr": ""}
-
-    def execute_create_discussion(
-        self, owner: str, repo: str, title: str, body: str, category: str, pat: str
-    ) -> dict:
-        """Discussion を作成"""
-        repo_id = self.get_repository_id(owner, repo, pat)
-        category_id = self.get_discussion_category_id(owner, repo, category, pat)
-
-        mutation = """
-        mutation($repositoryId: ID!, $categoryId: ID!, $title: String!, $body: String!) {
-            createDiscussion(input: {repositoryId: $repositoryId, categoryId: $categoryId, title: $title, body: $body}) {
-                discussion {
-                    number
-                    url
-                }
-            }
-        }
-        """
-        variables = {
-            "repositoryId": repo_id,
-            "categoryId": category_id,
-            "title": title,
-            "body": body,
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-        d = result["data"]["createDiscussion"]["discussion"]
-
-        return {
-            "exit_code": 0,
-            "stdout": d["url"],
-            "stderr": f"Created discussion #{d['number']}"
-        }
-
-    def execute_update_discussion(
-        self, owner: str, repo: str, number: int, title: str | None, body: str | None, pat: str
-    ) -> dict:
-        """Discussion を更新"""
-        discussion_id = self.get_discussion_node_id(owner, repo, number, pat)
-
-        mutation = """
-        mutation($discussionId: ID!, $title: String, $body: String) {
-            updateDiscussion(input: {discussionId: $discussionId, title: $title, body: $body}) {
-                discussion {
-                    number
-                    url
-                }
-            }
-        }
-        """
-        variables = {
-            "discussionId": discussion_id,
-            "title": title,
-            "body": body,
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-        d = result["data"]["updateDiscussion"]["discussion"]
-
-        return {
-            "exit_code": 0,
-            "stdout": d["url"],
-            "stderr": f"Updated discussion #{d['number']}"
-        }
-
-    def execute_add_discussion_comment(
-        self, owner: str, repo: str, number: int, body: str, reply_to: str | None, pat: str
-    ) -> dict:
-        """Discussion にコメントを追加"""
-        discussion_id = self.get_discussion_node_id(owner, repo, number, pat)
-
-        mutation = """
-        mutation($discussionId: ID!, $body: String!, $replyToId: ID) {
-            addDiscussionComment(input: {discussionId: $discussionId, body: $body, replyToId: $replyToId}) {
-                comment {
-                    id
-                    url
-                }
-            }
-        }
-        """
-        variables = {
-            "discussionId": discussion_id,
-            "body": body,
-            "replyToId": reply_to,
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-        c = result["data"]["addDiscussionComment"]["comment"]
-
-        return {
-            "exit_code": 0,
-            "stdout": c["url"],
-            "stderr": f"Added comment {c['id']}"
-        }
-
-    def execute_update_discussion_comment(
-        self, comment_id: str, body: str, pat: str
-    ) -> dict:
-        """Discussion コメントを更新"""
-        mutation = """
-        mutation($commentId: ID!, $body: String!) {
-            updateDiscussionComment(input: {commentId: $commentId, body: $body}) {
-                comment {
-                    id
-                    url
-                }
-            }
-        }
-        """
-        variables = {
-            "commentId": comment_id,
-            "body": body,
-        }
-        result = self.execute_graphql(mutation, variables, pat)
-        c = result["data"]["updateDiscussionComment"]["comment"]
-
-        return {
-            "exit_code": 0,
-            "stdout": c["url"],
-            "stderr": f"Updated comment {c['id']}"
-        }
-
-    def do_GET(self):
-        self.route_request("GET")
-
-    def do_POST(self):
-        self.route_request("POST")
-
-    def do_PUT(self):
-        self.route_request("PUT")
-
-    def do_PATCH(self):
-        self.route_request("PATCH")
-
-    def do_DELETE(self):
-        self.route_request("DELETE")
-
-    def do_OPTIONS(self):
-        """CORS preflight"""
-        self.send_response(200)
-        self.send_header("Access-Control-Allow-Origin", "*")
-        self.send_header("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
-        self.send_header("Access-Control-Allow-Headers", "Content-Type")
-        self.end_headers()
-
-
-# =============================================================================
-# メイン
-# =============================================================================
-
-def main():
-    parser = argparse.ArgumentParser(description="GitHub Proxy")
-    parser.add_argument(
-        "--port", type=int, default=DEFAULT_PORT, help=f"Port to listen on (default: {DEFAULT_PORT})"
-    )
-    parser.add_argument(
-        "--config",
-        type=Path,
-        default=DEFAULT_CONFIG_PATH,
-        help=f"Config file path (default: {DEFAULT_CONFIG_PATH})",
-    )
-    args = parser.parse_args()
-
-    config = load_config(args.config)
-    GitHubProxyHandler.config = config
-
-    server = HTTPServer(("0.0.0.0", args.port), GitHubProxyHandler)
-    print(f"GitHub Proxy listening on http://0.0.0.0:{args.port}")
-    print(f"Config: {args.config}")
-
-    print(f"\nSupported actions:")
-    for category, actions in ACTION_CATEGORIES.items():
-        print(f"  {category}: {', '.join(a.split(':')[1] for a in actions)}")
-
-    print(f"\nPolicy rules: {len(config['rules'])}")
-    for i, rule in enumerate(config["rules"]):
-        effect = rule["effect"].upper()
-        actions = ", ".join(rule["actions"])
-        repos = ", ".join(rule["repos"])
-        print(f"  [{i}] {effect}: {actions} on {repos}")
-
-    print("\nPress Ctrl+C to stop")
-
-    try:
-        server.serve_forever()
-    except KeyboardInterrupt:
-        print("\nShutting down...")
-        server.shutdown()
-
+from fgp.server import main
 
 if __name__ == "__main__":
     main()
